requirements.txt
ansi2html==1.9.1
beautifulsoup4==4.12.3
discord.py==2.3.2
python-dotenv==1.0.1
